refactor(event box): search content에서 사용하던 event box를 공통 컴포넌트로 만들어 total content 에서도 사용 #317
